refactor(core): Requested Changes
Began creation of a helper class to reduce duplicate code among classes in the queue/implementation/blocks package. Also refactored duplicate code in recently added if/else blocks into private methods.

+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/queue/implementation/blocks/IntGetBlocks.java
@@ -8,6 +8,9 @@ import com.sk89q.worldedit.world.block.BlockTypesCache;
 
 import java.util.Arrays;
 
+/**
+ * @since TODO
+ */
 public abstract class IntGetBlocks extends IntBlocks implements IChunkGet {
 
     /**
@@ -36,7 +39,7 @@ public abstract class IntGetBlocks extends IntBlocks implements IChunkGet {
         if (data == null) {
             data = new int[4096];
         }
-        Arrays.fill(data, (int) BlockTypesCache.ReservedIDs.AIR);
+        Arrays.fill(data, BlockTypesCache.ReservedIDs.AIR);
         return data;
     }
